DataPipelines/bbbot1_pipeline/requirements.txt
# BBBot1 Pipeline & MLFlow Dependencies
# Core dependencies for data pipeline and MLFlow tracking

# dbt - Data transformation
dbt-core>=1.7.0
dbt-mysql>=1.7.0
dbt-snowflake>=1.7.0

# MLFlow and database
mlflow>=2.8.0
pymysql>=1.1.0
sqlalchemy>=1.4.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.10.0

# API clients
yfinance>=0.2.0
requests>=2.28.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Airflow (if running locally)
apache-airflow>=2.8.0

# Web scraping (for Yahoo portfolio scraping)
beautifulsoup4>=4.11.0
lxml>=4.9.0
